Inline configuration and a bugfix of GAC reference
When writing a simple library in Linqpad we will often need to test it
in the Main code, which may uses some specific references only needed
for the Main code, to remove them when building the library, we can now
embed a hashtable in special comment such as `// lpmake @{ExeOnly=...`
to describe which references/nuget/namespaces are for the Main code,
thus they can be removed when building the library.

Fixed a bug when adding GAC reference, ensure to use full name so that a
specific version of a GAC assembly can be used.